Alpine: added explicit non-virtual dependencies.
Due to a change in behaviour in apk-tools 2.12.10 [1], virtual provides are no
longer enough to satisfy a dependency even if they are provided by only one
real package.  A workaround seems to be to depend on a real package name as
well as on a virtual one.

For modules, dependant on NDK, added an explicit non-versioned dependency.

[1] https://gitlab.alpinelinux.org/alpine/apk-tools/-/commit/3b013f458225c2ad8a0d96ec3eb3dde2533e0312