njs: explicitely enable quickjs and link with libatomic
On some architectures without atomics they are implemented with libatomic and
are needed when linking quickjs to a resulting njs cli.

Currently the only known architecture is arm32v5, but there might be more.
Adding -latomic unconditionally is harmless since linker will decide whether to
use it or not due to -Wl,-as-needed flag passed during the build.