fix: Handle intermittent network errors
Intermittent network errors may happen when a module's source code is
downloaded. That happened to me recently (an unexpected 502 from GitHub)
which resulted in a missing module in my Docker image (that was
partially my fault, as the build process should stop in case of any
error).

This PR implements automatic retries for common transient network
hiccups.